refactoring for moving some implementations into model/ and validator/
