+++ management-api/post_product_test.go
@@ -5,6 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/future-architect/apidoor/managementapi"
+	"github.com/future-architect/apidoor/managementapi/model"
+	"github.com/future-architect/apidoor/managementapi/validator"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"io"
@@ -35,8 +37,8 @@ func TestPostProduct(t *testing.T) {
 	}()
 
 	// set up api info db
-	apiInfoList := managementapi.APIInfoList{
-		List: []managementapi.APIInfo{
+	apiInfoList := model.APIInfoList{
+		List: []model.APIInfo{
 			{
 				Name:        "info1",
 				Source:      "info1 company",
@@ -71,22 +73,22 @@ func TestPostProduct(t *testing.T) {
 	tests := []struct {
 		name              string
 		contentType       string
-		req               managementapi.PostProductReq
+		req               model.PostProductReq
 		wantHTTPStatus    int
 		wantResp          interface{}
-		wantProductRecord []managementapi.Product
+		wantProductRecord []model.Product
 		wantContentRecord []productAPIContent
 	}{
 		{
 			name:        "post product containing multiple APIs",
 			contentType: "application/json",
-			req: managementapi.PostProductReq{
+			req: model.PostProductReq{
 				Name:        "product1",
 				DisplayName: "product 1",
 				Source:      "company 1",
 				Description: "product 1 has two APIs",
 				Thumbnail:   "http://example.com/test",
-				Contents: []managementapi.APIContent{
+				Contents: []model.APIContent{
 					{
 						ID:          apiInfoList.List[0].ID,
 						Description: "first api",
@@ -100,7 +102,7 @@ func TestPostProduct(t *testing.T) {
 			},
 			wantHTTPStatus: http.StatusCreated,
 			wantResp:       "Created",
-			wantProductRecord: []managementapi.Product{
+			wantProductRecord: []model.Product{
 				{
 					Name:        "product1",
 					DisplayName: "product 1",
@@ -123,12 +125,12 @@ func TestPostProduct(t *testing.T) {
 		{
 			name:        "post product when some optional fields are omitted",
 			contentType: "application/json",
-			req: managementapi.PostProductReq{
+			req: model.PostProductReq{
 				Name:        "product2",
 				Source:      "company 2",
 				Description: "product 2 has one API",
 				Thumbnail:   "http://example.com/test",
-				Contents: []managementapi.APIContent{
+				Contents: []model.APIContent{
 					{
 						ID: apiInfoList.List[0].ID,
 					},
@@ -137,7 +139,7 @@ func TestPostProduct(t *testing.T) {
 			},
 			wantHTTPStatus: http.StatusCreated,
 			wantResp:       "Created",
-			wantProductRecord: []managementapi.Product{
+			wantProductRecord: []model.Product{
 				{
 					Name:            "product2",
 					Source:          "company 2",
@@ -155,17 +157,17 @@ func TestPostProduct(t *testing.T) {
 		{
 			name:        "post product containing no API is allowed",
 			contentType: "application/json",
-			req: managementapi.PostProductReq{
+			req: model.PostProductReq{
 				Name:        "product3",
 				Source:      "company 3",
 				Description: "product 3 has no API",
 				Thumbnail:   "http://example.com/test",
-				Contents:    []managementapi.APIContent{},
+				Contents:    []model.APIContent{},
 				IsAvailable: false,
 			},
 			wantHTTPStatus: http.StatusCreated,
 			wantResp:       "Created",
-			wantProductRecord: []managementapi.Product{
+			wantProductRecord: []model.Product{
 				{
 					Name:        "product3",
 					Source:      "company 3",
@@ -178,13 +180,13 @@ func TestPostProduct(t *testing.T) {
 		{
 			name:        "api id not existing is contained",
 			contentType: "application/json",
-			req: managementapi.PostProductReq{
+			req: model.PostProductReq{
 				Name:        "product91",
 				DisplayName: "product 91",
 				Source:      "company 91",
 				Description: "product 91 is wrong",
 				Thumbnail:   "http://example.com/test",
-				Contents: []managementapi.APIContent{
+				Contents: []model.APIContent{
 					{
 						ID:          notExistID,
 						Description: "api not existing",
@@ -193,28 +195,28 @@ func TestPostProduct(t *testing.T) {
 				IsAvailable: false,
 			},
 			wantHTTPStatus: http.StatusBadRequest,
-			wantResp: managementapi.BadRequestResp{
+			wantResp: validator.BadRequestResp{
 				Message:          fmt.Sprintf("api_id %d does not exist", notExistID),
 				ValidationErrors: nil,
 			},
-			wantProductRecord: []managementapi.Product{},
+			wantProductRecord: []model.Product{},
 			wantContentRecord: []productAPIContent{},
 		},
 		{
 			name:        "input validation is failed",
 			contentType: "application/json",
-			req: managementapi.PostProductReq{
+			req: model.PostProductReq{
 				Name:        "product 92",
 				DisplayName: "product 92",
 				Description: "product 92 is wrong",
 				Thumbnail:   "http://example.com/test",
-				Contents:    []managementapi.APIContent{},
+				Contents:    []model.APIContent{},
 				IsAvailable: false,
 			},
 			wantHTTPStatus: http.StatusBadRequest,
-			wantResp: managementapi.BadRequestResp{
+			wantResp: validator.BadRequestResp{
 				Message: "input validation error",
-				ValidationErrors: &managementapi.ValidationErrors{
+				ValidationErrors: &validator.ValidationErrors{
 					{
 						Field:          "name",
 						ConstraintType: "alphanum",
@@ -229,7 +231,7 @@ func TestPostProduct(t *testing.T) {
 					},
 				},
 			},
-			wantProductRecord: []managementapi.Product{},
+			wantProductRecord: []model.Product{},
 			wantContentRecord: []productAPIContent{},
 		},
 	}
@@ -266,8 +268,8 @@ func TestPostProduct(t *testing.T) {
 				if tt.wantResp != string(resp) {
 					t.Errorf("wrong reponse body: got %s, want %s", resp, tt.wantResp)
 				}
-			case managementapi.BadRequestResp:
-				want := tt.wantResp.(managementapi.BadRequestResp)
+			case validator.BadRequestResp:
+				want := tt.wantResp.(validator.BadRequestResp)
 				testBadRequestResp(t, &want, resp)
 			default:
 				t.Errorf("type of wantResp is not unsupported")
@@ -281,9 +283,9 @@ func TestPostProduct(t *testing.T) {
 				t.Errorf("db get api info error: %v", err)
 				return
 			}
-			productList := make([]managementapi.Product, 0)
+			productList := make([]model.Product, 0)
 			for rows.Next() {
-				var row managementapi.Product
+				var row model.Product
 				if err := rows.StructScan(&row); err != nil {
 					t.Errorf("reading row error: %v", err)
 					return
@@ -291,7 +293,7 @@ func TestPostProduct(t *testing.T) {
 				productList = append(productList, row)
 			}
 			if diff := cmp.Diff(tt.wantProductRecord, productList,
-				cmpopts.IgnoreFields(managementapi.Product{}, "ID")); diff != "" {
+				cmpopts.IgnoreFields(model.Product{}, "ID")); diff != "" {
 				t.Errorf("db get list of product responce differs:\n %v", diff)
 			}
