+++ management-api/post_user_test.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/future-architect/apidoor/managementapi"
+	"github.com/future-architect/apidoor/managementapi/model"
+	"github.com/future-architect/apidoor/managementapi/validator"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"io"
@@ -23,16 +25,16 @@ func TestPostUser(t *testing.T) {
 	tests := []struct {
 		name               string
 		contentType        string
-		req                managementapi.PostUserReq
+		req                model.PostUserReq
 		wantHttpStatus     int
-		wantBadRequestResp *managementapi.BadRequestResp
+		wantBadRequestResp *validator.BadRequestResp
 		//wantRecord は期待されるDB作成レコードの値、idは比較対象外
-		wantRecords []managementapi.User
+		wantRecords []model.User
 	}{
 		{
 			name:        "正常に登録できる",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "user",
 				EmailAddress: "test00@example.com",
 				Password:     "password",
@@ -40,7 +42,7 @@ func TestPostUser(t *testing.T) {
 			},
 			wantHttpStatus:     http.StatusCreated,
 			wantBadRequestResp: nil,
-			wantRecords: []managementapi.User{
+			wantRecords: []model.User{
 				{
 					AccountID:      "user",
 					EmailAddress:   "test00@example.com",
@@ -52,7 +54,7 @@ func TestPostUser(t *testing.T) {
 		{
 			name:        "パスワードに記号が含まれており、正常に登録できる",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "user1",
 				EmailAddress: "test01@example.com",
 				Password:     "p@ss12Word",
@@ -60,7 +62,7 @@ func TestPostUser(t *testing.T) {
 			},
 			wantHttpStatus:     http.StatusCreated,
 			wantBadRequestResp: nil,
-			wantRecords: []managementapi.User{
+			wantRecords: []model.User{
 				{
 					AccountID:      "user1",
 					EmailAddress:   "test01@example.com",
@@ -72,7 +74,7 @@ func TestPostUser(t *testing.T) {
 		{
 			name:        "名前が設定されていなくても、正常に登録できる",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "user2",
 				EmailAddress: "test02@example.com",
 				Password:     "password",
@@ -80,7 +82,7 @@ func TestPostUser(t *testing.T) {
 			},
 			wantHttpStatus:     http.StatusCreated,
 			wantBadRequestResp: nil,
-			wantRecords: []managementapi.User{
+			wantRecords: []model.User{
 				{
 					AccountID:      "user2",
 					EmailAddress:   "test02@example.com",
@@ -92,16 +94,16 @@ func TestPostUser(t *testing.T) {
 		{
 			name:        "必須項目に空欄があるとき、登録できない",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "",
 				EmailAddress: "test03@example.com",
 				Password:     "password",
 				Name:         "full name",
 			},
 			wantHttpStatus: http.StatusBadRequest,
-			wantBadRequestResp: &managementapi.BadRequestResp{
+			wantBadRequestResp: &validator.BadRequestResp{
 				Message: "input validation error",
-				ValidationErrors: &managementapi.ValidationErrors{
+				ValidationErrors: &validator.ValidationErrors{
 					{
 						Field:          "account_id",
 						ConstraintType: "required",
@@ -110,21 +112,21 @@ func TestPostUser(t *testing.T) {
 					},
 				},
 			},
-			wantRecords: []managementapi.User{},
+			wantRecords: []model.User{},
 		},
 		{
 			name:        "account_idにprintable ascii以外の文字が含まれていたとき、登録できない",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "userユーザー",
 				EmailAddress: "test04@example.com",
 				Password:     "password",
 				Name:         "full name",
 			},
 			wantHttpStatus: http.StatusBadRequest,
-			wantBadRequestResp: &managementapi.BadRequestResp{
+			wantBadRequestResp: &validator.BadRequestResp{
 				Message: "input validation error",
-				ValidationErrors: &managementapi.ValidationErrors{
+				ValidationErrors: &validator.ValidationErrors{
 					{
 						Field:          "account_id",
 						ConstraintType: "printascii",
@@ -133,21 +135,21 @@ func TestPostUser(t *testing.T) {
 					},
 				},
 			},
-			wantRecords: []managementapi.User{},
+			wantRecords: []model.User{},
 		},
 		{
 			name:        "email_addressの文字列がメールアドレスとして不正であるとき、登録できない",
 			contentType: "application/json",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "user",
 				EmailAddress: "test05.@example.com",
 				Password:     "password",
 				Name:         "full name",
 			},
 			wantHttpStatus: http.StatusBadRequest,
-			wantBadRequestResp: &managementapi.BadRequestResp{
+			wantBadRequestResp: &validator.BadRequestResp{
 				Message: "input validation error",
-				ValidationErrors: &managementapi.ValidationErrors{
+				ValidationErrors: &validator.ValidationErrors{
 					{
 						Field:          "email_address",
 						ConstraintType: "email",
@@ -156,22 +158,22 @@ func TestPostUser(t *testing.T) {
 					},
 				},
 			},
-			wantRecords: []managementapi.User{},
+			wantRecords: []model.User{},
 		},
 		{
 			name:        "Content-Typeがapplication/json以外であるとき、登録できない",
 			contentType: "text/plain",
-			req: managementapi.PostUserReq{
+			req: model.PostUserReq{
 				AccountID:    "user",
 				EmailAddress: "test06@example.com",
 				Password:     "password",
 				Name:         "full name",
 			},
 			wantHttpStatus: http.StatusBadRequest,
-			wantBadRequestResp: &managementapi.BadRequestResp{
+			wantBadRequestResp: &validator.BadRequestResp{
 				Message: `unexpected request Content-Type, it must be "application/json"`,
 			},
-			wantRecords: []managementapi.User{},
+			wantRecords: []model.User{},
 		},
 	}
 
@@ -216,9 +218,9 @@ func TestPostUser(t *testing.T) {
 				return
 			}
 
-			list := []managementapi.User{}
+			list := []model.User{}
 			for rows.Next() {
-				var row managementapi.User
+				var row model.User
 
 				if err := rows.StructScan(&row); err != nil {
 					t.Errorf("reading row error: %v", err)
@@ -229,7 +231,7 @@ func TestPostUser(t *testing.T) {
 			}
 
 			if diff := cmp.Diff(tt.wantRecords, list,
-				cmpopts.IgnoreFields(managementapi.User{}, "ID", "LoginPasswordHash",
+				cmpopts.IgnoreFields(model.User{}, "ID", "LoginPasswordHash",
 					"CreatedAt", "UpdatedAt")); diff != "" {
 				t.Errorf("db get users responce differs:\n %v", diff)
 			}
@@ -250,8 +252,8 @@ func TestPostUser(t *testing.T) {
 
 }
 
-func testBadRequestResp(t *testing.T, want *managementapi.BadRequestResp, got []byte) {
-	var gotBody managementapi.BadRequestResp
+func testBadRequestResp(t *testing.T, want *validator.BadRequestResp, got []byte) {
+	var gotBody validator.BadRequestResp
 	if err := json.Unmarshal(got, &gotBody); err != nil {
 		t.Errorf("parsing body as BadRequestResp failed: %v", err)
 		return
